Fix the issue for cursor 'not-allowed'
When I tried to make the app work only when the button is both clicked
and over a grid item, and I started drawing, it often displayed the
cursor as 'not-allowed' and broke the logic.

Turns out the solution is to adjust the CSS with the following props on
the grid items:
  user-select: none

This fixed the issue and now I can free draw.